refactor(chat): move ToolExecutor and optimize avatar generation
- Move ToolExecutor.tsx from tools/ to chat/ directory
- Update import path in MyRuntimeProvider accordingly
- Optimize gimli avatar generation by caching result
- Avoid repeated function calls on each render
